moved CartService::addProductToCart to Cart::addProduct  - removed CartService
